No Hassel - HCI Project
Application name: “No-Hassel”
Description: "No-Hassel" is a mobile application designed to help individuals with visual and speech impairments overcome barriers in daily communication and access to information. The application leverages image recognition and text-to-speech technology to assist individuals with visual impairments in identifying and reading text on objects, signs, and screens. Additionally, the application provides speech-to-text and text-to-speech capabilities, allowing individuals with speech impairments to communicate with others in real-time. The goal of No-Hassel is to provide a convenient and accessible solution for individuals with disabilities, enhancing their independence and quality of life."